Update README.md
amazon freertos as an external library. =amazon Freertos as an external library.
cmake --build build --target flash =CMake --build --target flash
If you changed some configuartion and run into this, usually it's  because of missing dependent libraries/demos. To fix it, first find out which CMake target the missing functions belong to, normally it should be in the `CMakeLists.txt` closest to the source file where the functions are defined. Then add the CMake target to your application dependency list using the `target_link_libraries` command in the root level `CMakeLists.txt`. For example,   ==
If you changed some configuration and run into this, usually it's because of missing dependent libraries/demos. To fix it, first, find out which CMake target the missing functions belong to, normally it should be in the `CMakeLists.txt` closest to the source file where the functions are defined. Then add the CMake target to your application dependency list using the `target_link_libraries` command in the root level `CMakeLists.txt`. For example,